Added/Updated Sensors
Added the temperature sensors in order to tell the user how hot/cold it is within the device. I also improved the supersonic sensors in order to instantly read the response from a users hand as well as other objects around. And added another sensor in the back of the robot to prevent it from backing into a wall.