Logic for connection purposes and RSSI value intake
Due to the server being slow to connect when the server is innaccessible I had to find a way to allow smooth operations when obtaining RSSI values without the influence of the client connection being an issue. I was able to create logic in which I needed to check the connection upon the server 3 times before continuously obtaining the rssi values for 10 seconds then looking for a connection to the server.