Added Privacy and Data page & began login
Still need to figure out how to delete user data from backend
before that, need to be able to create and login

also added a user icon on far right
currently sends username and password as clear text

also, when in the text field, if I hit s, it jumps to settings. Incredible.