added the button functionality and refresh corrections
The buttons and the obtaining of the temperature information works now. I am able to send the temperature data up to the server and send the direction information back to the robot. When it comes to refreshing the pages it didn't work correctly. So now I was able to add a browserRouter which allows for the refreshment of the pages.