Increased the refresh rate speed on the SPI TFT display
So apparently it does not need a identifier, It just needed a frequency in which the display board should refresh. The board was not set to a frequency since we are using an Arduino Uno R4 board which is completely different to other microcontrollers. So I set the refresh rate to 8 Mhz and it was able to project information onto the screen.