Added Templates
I added templates to make my code more concise in what I need to do. When it comes to obtaining the enum information through the unordered map it improves the standalone code and makes the functionality of the buttons more precise. And the loading times because of the map is improved.