Keyboard functionality begins
I began adding the keyboard functionality, and began testing different ways in adding the functionality. I found one that I plan to stick with till the end. And I believe it's useful enough. HOWEVER, Securing the users ssid and password safely is starting to become a rising concern. Sure it is stored in the EEPROM but if someone were to connect to the device they could manage to obtain the ssid and password. So while I am working on the keyboard I am looking into ways in how I can encrypt the password and ssid while it is in the EEPROM.