Backend implemented
Implemented the password/ssid storage where when you input an ssid and password it will be saved within the eeprom. After clicking the save button, the information will be saved.

Minor Bug: When inputting a new ssid or password during the connected phase, it sticks with the previous password and ssid for some reason. Will need to debug more tomorrow. But for now it works as it stands.