Removed specific workflows/ updated readme
I updated the readme to include necessary techniques that are needed to develop
on this repository, unit testing will be necessary no matter what. Workflows will be replaced with robotic based workflows. Arduino Code most likely will be migrated at somepoint due to the fact that It would be best to deal with new code compared to
Legacy code.